# back/docs/swagger.yaml
basePath: /api/v1
definitions:
  errors.ErrorResponse:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
...
  models.CreateTagPayload:
    properties:
      label:
        type: string
    required:
    - label
    type: object
  models.Game:
    properties:
      created_at:
        type: string
      id:
        type: integer
      media:
        $ref: '#/definitions/models.Media'
      media_id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
    type: object
...
  models.Match:
    properties:
      created_at:
        type: string
      endTime:
        type: string
      id:
        type: integer
      matchPosition:
        type: integer
      playerOne:
        $ref: '#/definitions/models.User'
      playerOneID:
        type: integer
      playerTwo:
        $ref: '#/definitions/models.User'
      playerTwoID:
        type: integer
      scores:
        items:
          $ref: '#/definitions/models.Score'
        type: array
      startTime:
        type: string
      status:
        type: string
      tournament:
        $ref: '#/definitions/models.Tournament'
      tournamentID:
        type: integer
      tournamentStep:
        $ref: '#/definitions/models.TournamentStep'
      tournamentStepID:
        type: integer
      updated_at:
        type: string
      winner:
        $ref: '#/definitions/models.User'
      winnerID:
        type: integer
    required:
    - status
    type: object
  models.Media:
    properties:
      created_at:
        type: string
      file_extension:
        type: string
      file_name:
        type: string
      id:
        type: integer
      updated_at:
        type: string
    type: object
...
  models.Score:
    properties:
      created_at:
        type: string
      id:
        type: integer
      match:
        $ref: '#/definitions/models.Match'
      matchID:
        type: integer
      player:
        $ref: '#/definitions/models.User'
      playerID:
        type: integer
      score:
        type: integer
      updated_at:
        type: string
    type: object
  models.Tag:
    properties:
      created_at:
        type: string
      games:
        items:
          $ref: '#/definitions/models.Game'
        type: array
      id:
        type: integer
      name:
        type: string
      tournaments:
        items:
          $ref: '#/definitions/models.Tournament'
        type: array
      updated_at:
        type: string
    required:
    - name
    type: object
  models.Tournament:
    properties:
      created_at:
        type: string
      description:
        type: string
      end_date:
        type: string
      game:
        $ref: '#/definitions/models.Game'
      game_id:
        type: integer
      id:
        type: integer
      location:
        type: string
      maxPlayers:
        example: 32
        type: integer
      media:
        $ref: '#/definitions/models.Media'
      media_id:
        type: integer
      name:
        type: string
      organizer_id:
        type: integer
      rounds:
        type: integer
      start_date:
        type: string
      status:
        type: string
      tags:
        items:
          $ref: '#/definitions/models.Tag'
        type: array
      tournament_steps:
        items:
          $ref: '#/definitions/models.TournamentStep'
        type: array
      updated_at:
        type: string
      user:
        $ref: '#/definitions/models.User'
      users:
        items:
          $ref: '#/definitions/models.User'
        type: array
    required:
    - description
    - end_date
    - game_id
    - maxPlayers
    - name
    - rounds
    - start_date
    type: object
  models.TournamentStep:
    properties:
      created_at:
        type: string
      id:
        type: integer
      matches:
        items:
          $ref: '#/definitions/models.Match'
        type: array
      name:
        type: string
      sequence:
        type: integer
      tournamentID:
        type: integer
      updated_at:
        type: string
    type: object
  models.User:
    properties:
      address:
        type: string
      country:
        type: string
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      is_verified:
        type: boolean
      matches:
        items:
          $ref: '#/definitions/models.Match'
        type: array
      media:
        $ref: '#/definitions/models.Media'
      media_id:
        type: integer
      password:
        minLength: 6
        type: string
      phone:
        type: string
      role:
        type: string
      tournaments:
        items:
          $ref: '#/definitions/models.Tournament'
        type: array
      updated_at:
        type: string
      username:
        type: string
    required:
    - password
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server celler server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Swagger Example API
  version: "1.0"
paths:
...
  /matchs:
    get:
      consumes:
      - application/json
      description: Get all matchs.
      parameters:
      - description: Search by userID
        in: query
        name: UserID
        type: integer
      - description: Search by TournamentID
        in: query
        name: TournamentID
        type: integer
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get all matchs
      tags:
      - Match
  /matchs/{id}:
    put:
      consumes:
      - application/json
      description: Update a match with the given ID.
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Match ID
        in: path
        name: id
        required: true
        type: string
      - description: Match data
        in: body
        name: match
        required: true
        schema:
          $ref: '#/definitions/models.Match'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update a match
      tags:
      - Match
  /matchs/{matchId}:
    get:
      consumes:
      - application/json
      description: Get a match with the given ID.
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Match ID
        in: path
        name: matchId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Match'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      security:
      - BearerAuth®: []
      summary: Get a match
      tags:
      - Match
  /matchs/update/score:
    post:
      consumes:
      - application/json
      description: Update score.
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Match ID
        in: formData
        name: matchId
        required: true
        type: string
      - description: User ID
        in: formData
        name: userId
        required: true
        type: string
      - description: Score
        in: formData
        name: score
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update score
      tags:
      - Match
...
  /tags:
    get:
      consumes:
      - application/json
      description: Get all tags.
      parameters:
      - description: Add trendy games to the response
        in: query
        name: WithTrendy
        type: boolean
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get all tags
      tags:
      - Tag
    post:
      consumes:
      - application/json
      description: Create a tag
      parameters:
      - description: Tag object that needs to be created
        in: body
        name: tag
        required: true
        schema:
          $ref: '#/definitions/models.CreateTagPayload'
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Tag'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a tag
      tags:
      - Tag
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"